all minor updates

+++ src/main/java/serviceImpl/AdminServiceImpl.java
@@ -1,5 +1,6 @@
 package serviceImpl;
 
+import enums.Role;
 import model.User;
 import enums.ResponseStatus;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +24,7 @@ public class AdminServiceImpl implements AdminService {
         try {
             Object userPresentOrNot = userService.checkUserPrentOrNot(user.getEmail()).getResponseObject();
             if (Objects.isNull(userPresentOrNot)) {
-                response =  userService.addUser(user);
+                response = userService.addUser(user);
             } else {
                 response = Response.builder().message("user already present...").statusCode(ResponseStatus.Error).build();
             }
@@ -39,11 +40,14 @@ public class AdminServiceImpl implements AdminService {
         Response response = new Response();
         try {
             Object user = userService.fetchUserByEmail(email).getResponseObject();
-            if(Objects.isNull(user)){
+            if (Objects.nonNull(user) && user instanceof User userData) {
+                if (userData.getRole() != Role.user) {
+                    response = userService.removeUser(userData);
+                } else {
+                    response = Response.builder().responseObject(null).message("user not found").statusCode(ResponseStatus.Error).build();
+                }
+            } else {
                 response = Response.builder().responseObject(null).message("user not found").statusCode(ResponseStatus.Error).build();
-            }
-            if (user instanceof User userData) {
-               response = userService.removeUser(userData);
             }
         } catch (Exception e) {
             response = Response.builder().message("something went wrong try again...").statusCode(ResponseStatus.Error).build();
@@ -56,7 +60,7 @@ public class AdminServiceImpl implements AdminService {
     public Response fetchUser() {
         Response response;
         try {
-            response =  userService.fetchUser();
+            response = userService.fetchUser();
         } catch (Exception e) {
             response = Response.builder().message("Something went wrong try again...").statusCode(ResponseStatus.Error).build();
             log.error("Something went wrong...");
